feat: New structure for productscarousel
